bugfix: results with resultAcquisitionError break the view
Sometimes cucumber errors out and the framework collects details why
and then adds that information to the JSON property "resultAcquisitionError".
When this happens, the cucumberResult property is null; the GUI assumed
it was an empty array. With this fix, we automatically make it an empty array
when it is null.

JIRA: #DEVOPS-375